Optimized RSN writing.
Previously, OUI and subtype were written separatly, which shouldn't have
been an issue, but somehow led to suboptimal codegen. They have now been
merge.

Along with that, the writing of the AKM, cipher suite and PMKID-list was
altered. Instead of requiring them to implement `IntoIterator`, they now
require `TryIntoCtx` instead, which is implemented by all containers and
`ReadIterator`.

Codegen is still not as good, as I would like, but this is a step in the
right direction.