Fixed match_frames!.
The first issue was, that previously scroll wasn't imported inside
match_frames!, which caused errors while compiling. This went unnoticed,
because scroll is a dep of this crate and was therefore available in all
tests.

The second issue was, that correctly matching management frames were
reported as not matching by GenericFrame::matches. This has since been
resolved.